+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - **Breaking Change**: in process substitution, internal job object no longer maintain created ``FD`` object
 - **Breaking Change**: ``break``, ``continue``, ``return`` expressions are no longer allowed in last-pipe
+- **Breaking Change**: when ``err-raise`` option is enabled, check pipeline exit status even if last-pipe
 - improve assertion error message of binary ``==`` and ``=~`` expression
     - now show left hand-side, right hand-side expression value